+++ src/database/redisDB.ts
@@ -1,0 +1,6 @@
+import { createClient } from 'redis'
+
+const redis = createClient();
+redis.on('error', (err: Error) => console.log('Redis Client Error', err));
+
+export { redis }