+++ src/utils/permission-required.test.ts
@@ -11,7 +11,7 @@ afterAll(() => {
 
 describe('permissionRequired', () => {
     it('should log the request route', () => {
-        const req = buildReq({ route: 'TEST_ROUTE' })
+        const req = buildReq({ url: 'TEST_ROUTE' })
         const res = buildRes()
         const next = buildNext()
         // mock the console.log function to supress the test logs
@@ -19,12 +19,12 @@ describe('permissionRequired', () => {
         jest.spyOn(global.console, 'log')
 
 
-        permissionRequired(req, res, next)
+        permissionRequired('code')(req, res, next)
 
         expect(next).toHaveBeenCalledWith(/* nothing */)
         expect(next).toHaveBeenCalledTimes(1)
 
-        expect(console.log).toHaveBeenCalledWith('=> TEST_ROUTE')
+        expect(console.log).toHaveBeenCalledWith('=> code:TEST_ROUTE')
         expect(console.log).toHaveBeenCalledTimes(1)
     })
 })