+++ src/utils/permission-required.ts
@@ -1,12 +1,13 @@
-import {NextFunction, Request, Response} from "express"
+import { NextFunction, Request, Response } from 'express'
 
 /**
  * permissionRequired checks whether the user has permission to access the route
  * being requested.
  * */
-const permissionRequired = (req: Request, res: Response, next: NextFunction) => {
-    console.log(`=> ${req.route}`)
-    next()
+const permissionRequired = (permissionCode: string) =>
+    (req: Request, res: Response, next: NextFunction) => {
+    console.log(`=> ${permissionCode}:${req.url}`)
+        next()
 }
 
 export {