+++ test/generate.ts
@@ -1,11 +1,30 @@
-import { FlightLog } from '../src/database/flight-log'
+import { FlightLog } from 'includes/flight-log'
 
-const mockFlightLog = (overrides: Partial<FlightLog> = {}) => {
+const mockFlightLog = (overrides: Partial<FlightLog> = {}): FlightLog => {
     return {
         userUUID: '54f94bbc-1394-4796-a5a0-7e847afbb813',
         date: new Date('2022-11-21T18:48:23.000Z'),
         type: 'A310',
         registration: 'ZXYIOU',
+        pilotInCommand: 'SELF',
+        details: 'HKG-CPT',
+        instrumentNavAids: 'ivl',
+        instrumentPlace: 'jvl',
+        instrumentActual: 1,
+        instrumentFSTD: 0,
+        instructorSE: 0,
+        instructorME: 0,
+        instructorFSTD: 0,
+        FSTD: 0,
+        engineType: 'single',
+        dayType: 'day',
+        dual: 0,
+        PIC: 0,
+        PICUS: 0,
+        copilot: 0,
+        dayLandings: 0,
+        nightLandings: 1,
+        remarks: 'weather conditions not ideal',
         ...overrides
     }
 }