+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.1] - 2022-11-21
+### Added
+- Some tests to get basic test coverage.
+- Database config:
+    - `database/setup.sql` file for all the database tables and extensions.
+    - `database/data.sql` file for some basic testing data.
+- Initial handler scaffolding.
+
 ## [0.0.0] - 2022-11-12
 ### Added
 - Basic setup of the project.